# docs/api/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AlertIncidentResponse": {
        "description": "API response model for an alert incident.",
        "properties": {
          "consecutive_violations": {
            "title": "Consecutive Violations",
            "type": "integer"
          },
          "fired_at": {
            "format": "date-time",
            "title": "Fired At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_value": {
            "title": "Last Value",
            "type": "number"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "rule_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Id"
          },
          "rule_name": {
            "title": "Rule Name",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "slo_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slo Name"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "threshold": {
            "title": "Threshold",
            "type": "number"
          }
        },
        "required": [
          "id",
          "rule_id",
          "rule_name",
          "severity",
          "status",
          "last_value",
          "threshold",
          "consecutive_violations",
          "fired_at",
          "resolved_at"
        ],
        "title": "AlertIncidentResponse",
        "type": "object"
      },
      "AlertMessageResponse": {
        "description": "Standardized response container.",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "default": "success",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "AlertMessageResponse",
        "type": "object"
      },
      "AlertRuleResponse": {
        "description": "API response model for an alert rule.",
        "properties": {
          "comparison": {
            "title": "Comparison",
            "type": "string"
          },
          "consecutive_cycles": {
            "title": "Consecutive Cycles",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_enabled": {
            "title": "Is Enabled",
            "type": "boolean"
          },
          "metric": {
            "title": "Metric",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "threshold": {
            "title": "Threshold",
            "type": "number"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "window": {
            "title": "Window",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "metric",
          "threshold",
          "comparison",
          "window",
          "severity",
          "consecutive_cycles",
          "is_enabled",
          "created_at",
          "updated_at"
        ],
        "title": "AlertRuleResponse",
        "type": "object"
      },
      "AlertRuleSchema": {
        "description": "Pydantic schema for creating or updating an alert rule.",
        "properties": {
          "comparison": {
            "default": "gt",
            "enum": [
              "gt",
              "lt"
            ],
            "title": "Comparison",
            "type": "string"
          },
          "consecutive_cycles": {
            "default": 1,
            "minimum": 1.0,
            "title": "Consecutive Cycles",
            "type": "integer"
          },
          "is_enabled": {
            "default": true,
            "title": "Is Enabled",
            "type": "boolean"
          },
          "metric": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Metric",
            "type": "string"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "severity": {
            "default": "warning",
            "enum": [
              "warning",
              "critical"
            ],
            "title": "Severity",
            "type": "string"
          },
          "slo_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slo Name"
          },
          "threshold": {
            "title": "Threshold",
            "type": "number"
          },
          "window": {
            "default": "15m",
            "enum": [
              "15m",
              "1h",
              "1d",
              "30d"
            ],
            "title": "Window",
            "type": "string"
          }
        },
        "required": [
          "name",
          "metric",
          "threshold"
        ],
        "title": "AlertRuleSchema",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "An Agentic Planetary Risk Intelligence Platform. Phase 1 — foundational API skeleton.",
    "title": "GaiaOS",
    "version": "0.5.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/alert-rules": {
      "get": {
        "description": "Returns all configured alerting rules. Requires ADMIN role.",
        "operationId": "list_alert_rules_api_v1_admin_alert_rules_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertRuleResponse"
                  },
                  "title": "Response List Alert Rules Api V1 Admin Alert Rules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List configured alert rules",
        "tags": [
          "v1",
          "Admin Alerts"
        ]
      },
      "post": {
        "description": "Upserts an alert rule by name. If a rule with the name exists, updates its configuration. Requires ADMIN role.",
        "operationId": "upsert_alert_rule_api_v1_admin_alert_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertRuleSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRuleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create or update an alert rule (idempotent upsert)",
        "tags": [
          "v1",
          "Admin Alerts"
        ]
      }
    },
    "/api/v1/admin/alert-rules/{rule_id}": {
      "delete": {
        "description": "Delete an alert rule by primary key ID.",
        "operationId": "delete_alert_rule_api_v1_admin_alert_rules__rule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertMessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Alert rule not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete an alert rule",
        "tags": [
          "v1",
          "Admin Alerts"
        ]
      }
    },
    "/api/v1/admin/alerts": {
      "get": {
        "description": "Returns current and historical alert incidents. Requires ADMIN role.",
        "operationId": "list_alert_incidents_api_v1_admin_alerts_get",
        "parameters": [
          {
            "description": "Filter by incident status",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "firing",
                    "resolved"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by incident status",
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertIncidentResponse"
                  },
                  "title": "Response List Alert Incidents Api V1 Admin Alerts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List alert incidents",
        "tags": [
          "v1",
          "Admin Alerts"
        ]
      }
    },
...
  }
}